Initial non-working version of bsd29fs

+++ ancient-src/bsd29/sys/sys2.c
@@ -216,7 +216,7 @@ close()
 {
 	register struct file *fp;
 	register struct a {
-		int	fdes;
+		int16_t	fdes;
 	} *uap;
 
 	uap = (struct a *)u.u_ap;
